// web/src/generated/api/models/businessVertical.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type BusinessVertical = typeof BusinessVertical[keyof typeof BusinessVertical];


export const BusinessVertical = {
  SALON: 'SALON',
  CLINIC: 'CLINIC',
  CONSULTANT: 'CONSULTANT',
  SPA_WELLNESS: 'SPA_WELLNESS',
  FITNESS_STUDIO: 'FITNESS_STUDIO',
  TUTORING: 'TUTORING',
  PET_GROOMING: 'PET_GROOMING',
} as const;