// web/src/generated/api/models/invitableMembershipRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type InvitableMembershipRole = typeof InvitableMembershipRole[keyof typeof InvitableMembershipRole];


export const InvitableMembershipRole = {
  MANAGER: 'MANAGER',
  STAFF: 'STAFF',
} as const;