// web/src/generated/api/models/membershipRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type MembershipRole = typeof MembershipRole[keyof typeof MembershipRole];


export const MembershipRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  STAFF: 'STAFF',
} as const;